pyproject.toml
[project]
name = "discord-live-bot"
version = "0.1.0"
description = "Discord bot for Bilibili live notifications"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
  "discord.py>=2.5.0,<3.0.0",
  "davey==0.1.0",
  "PyNaCl>=1.5.0,<2.0.0",
  "httpx>=0.23.0,<0.24.0",
  "python-dotenv>=1.0.0,<2.0.0",
  "loguru>=0.7.2,<0.8.0",
  "grpcio>=1.78.0,<2.0.0",
  "protobuf>=6.31.1,<7.0.0",
  "playwright>=1.54.0,<2.0.0",
  "streamlink>=8.2.0,<9.0.0",
]

[project.scripts]
discord-live-bot = "discord_live_bot.main:main"

[dependency-groups]
dev = [
  "pytest>=8.0.0,<9.0.0",
  "pytest-asyncio>=0.23.0,<1.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/discord_live_bot"]
